+++ harpy/commands/qc.py
@@ -20,6 +20,7 @@ from harpy.common.workflow import Workflow
 @click.option('-t', '--threads', panel = "Workflow Options", default = 4, show_default = True, type = click.IntRange(4,999, clamp = True), help = 'Number of threads to use')
 @click.option('-a', '--trim-adapters', panel = "Parameters", type = str, help = 'Detect and trim adapters')
 @click.option('-U','--unlinked', panel = "Parameters", is_flag = True, default = False, help = "Treat input data as not linked reads")
+@click.option('--clean', hidden = True, panel = "Workflow Options", type = str, help = 'Delete the log (`l`), .snakemake (`s`), and/or workflow (`w`) folders when done')
 @click.option('--container', panel = "Workflow Options",  is_flag = True, default = False, help = 'Use a container instead of conda', callback=container_ok)
 @click.option('--setup-only', panel = "Workflow Options",  is_flag = True, hidden = True, show_default = True, default = False, help = 'Setup the workflow and exit')
 @click.option('--hpc', panel = "Workflow Options",  type = HPCProfile(), help = 'HPC submission YAML configuration file')
@@ -27,7 +28,7 @@ from harpy.common.workflow import Workflow
 @click.option('--skip-reports', panel = "Workflow Options",  is_flag = True, default = False, help = 'Don\'t generate HTML reports')
 @click.option('--snakemake', panel = "Workflow Options", type = SnakemakeParams(), help = 'Additional Snakemake parameters, in quotes')
 @click.argument('inputs', required=True, type=FASTQfile(), nargs=-1)
-def qc(inputs, output_dir, unlinked, min_length, max_length, trim_adapters, deduplicate, extra_params, threads, snakemake, skip_reports, quiet, hpc, container, setup_only):
+def qc(inputs, output_dir, unlinked, min_length, max_length, trim_adapters, deduplicate, extra_params, threads, snakemake, skip_reports, quiet, hpc, clean, container, setup_only):
     """
     FASTQ adapter removal, quality filtering, etc.
 
@@ -46,7 +47,7 @@ def qc(inputs, output_dir, unlinked, min_length, max_length, trim_adapters, dedu
     - `-d` removes optical PCR duplicates
       - recommended to skip at this step in favor of barcode-assisted deduplication after alignment
     """
-    workflow = Workflow("qc", "qc.smk", output_dir, container, quiet)
+    workflow = Workflow("qc", "qc.smk", output_dir, container, clean, quiet)
     workflow.setup_snakemake(threads, hpc, snakemake)
     workflow.reports = ["qc_bx_stats.qmd"]
     workflow.conda = ["qc", "report"]